requirements.txt
amqp==5.0.6
APScheduler==3.7.0
asgiref==3.3.4
billiard==3.6.4.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.1.7
django-celery-results==2.0.1
django-ckeditor==6.0.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-import-export==2.5.0
django-js-asset==1.2.2
django-shortuuidfield==0.1.3
django-simpleui==2020.9.26
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
drf-yasg==1.20.0
et-xmlfile==1.0.1
gevent==21.1.2
greenlet==1.0.0
idna==2.10
importlib-metadata==4.0.0
inflection==0.5.1
install==1.3.4
itypes==1.2.0
Jinja2==2.11.3
kombu==5.0.2
MarkupPy==1.14
MarkupSafe==1.1.1
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.7
packaging==20.9
Pillow==8.2.0
prompt-toolkit==3.0.18
psycopg2==2.7.7
psycopg2-binary==2.8.6
PyJWT==2.0.1
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
shortuuid==1.0.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
tencentcloud-sdk-python==3.0.376
typing-extensions==3.7.4.3
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
xlrd==1.2.0
xlwt==1.3.0
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.4.0
